uint->size_t

+++ examples/test_serialize.cpp
@@ -62,7 +62,7 @@ int main(int argc, char* argv[]) {
     }
     char c = 0;
     
-    for(uint i = addon::immortal.get_index(); i < 100000; ++i) {
+    for(size_t i = addon::immortal.get_index(); i < 100000; ++i) {
         std::cin >> c;
         --data;
         std::cout << i << "  " << data << std::endl;
