+++ src/shift_region_class.hpp
@@ -18,7 +18,7 @@ namespace perimeter {
             std::ifstream in(filename);
             std::string temp;
             if(in.is_open()) {
-                uint consistent(0);
+                size_t consistent(0);
                 
                 stage1_.clear();
                 stage1_.push_back(std::vector<std::string>());
@@ -49,13 +49,13 @@ namespace perimeter {
             }
             convert_1_to_2();
         }
-        shift_region_class(uint const & H, uint const & L, double const & spacing): H_(H), L_(L), N_(2), grow_level_(2) {
-            stage2_ = std::vector<std::vector<std::vector<uint>>>(N_, std::vector<std::vector<uint>>(H_, std::vector<uint>(L_, 0)));
+        shift_region_class(size_t const & H, size_t const & L, double const & spacing): H_(H), L_(L), N_(2), grow_level_(2) {
+            stage2_ = std::vector<std::vector<std::vector<size_t>>>(N_, std::vector<std::vector<size_t>>(H_, std::vector<size_t>(L_, 0)));
             set_grow(std::vector<bond_type>(1, qmc::right));
             
-            uint grow_count = 0;
-            for(uint j = 0; j < 2; ++j) {
-                for(uint i = 0; i < H_; ++i) {
+            size_t grow_count = 0;
+            for(size_t j = 0; j < 2; ++j) {
+                for(size_t i = 0; i < H_; ++i) {
                     stage2_[1][i][j] = 1;
                     if(std::round((2.0 - spacing) * H_) > grow_count++)
                         stage2_[0][i][j] = 1;
@@ -64,23 +64,23 @@ namespace perimeter {
             convert_2_to_1();
         }
         
-        void print(uint flags = 1) const {
+        void print(size_t flags = 1) const {
             if((flags&1) == 1) {
                 std::cout << "--------stage1-graphical--------" << std::endl;
-                for(uint n = 0; n < stage1_.size(); ++n) {
+                for(size_t n = 0; n < stage1_.size(); ++n) {
                     std::cout << "state: " << n << std::endl;
-                    for(uint i = 0; i < stage1_[n].size(); ++i) {
+                    for(size_t i = 0; i < stage1_[n].size(); ++i) {
                         std::cout << "    " << stage1_[n][i] << std::endl;
                     }
                 }
             }
             if((flags&2) == 2) {
                 std::cout << "--------stage2-matrix--------" << std::endl;
-                for(uint n = 0; n < N_; ++n) {
+                for(size_t n = 0; n < N_; ++n) {
                     std::cout << "state: " << n << std::endl;
-                    for(uint i = 0; i < H_; ++i) {
+                    for(size_t i = 0; i < H_; ++i) {
                         std::cout << "    ";
-                        for(uint j = 0; j < L_; ++j) {
+                        for(size_t j = 0; j < L_; ++j) {
                             std::cout << (*this)(n, i, j) << " ";
                         }
                         std::cout << std::endl;
@@ -88,10 +88,10 @@ namespace perimeter {
                 }
             }
         }
-        uint operator()(uint const & n, uint const & i, uint const & j) const {
+        size_t operator()(size_t const & n, size_t const & i, size_t const & j) const {
             return stage2_[n][i][j];
         }
-        bool operator()(uint const & i, uint const & j) const { //backwards compatible
+        bool operator()(size_t const & i, size_t const & j) const { //backwards compatible
             assert(N_ == 1);
             return bool(stage2_[0][i][j]);
         }
@@ -101,12 +101,12 @@ namespace perimeter {
         void invert() {
             if(N_ != 1)
                 return;
-            for(uint i = 0; i < H_; ++i)
-                for(uint j = 0; j < L_; ++j)
+            for(size_t i = 0; i < H_; ++i)
+                for(size_t j = 0; j < L_; ++j)
                     stage2_[0][i][j] = !stage2_[0][i][j];
         }
         
-        uint & get_neighbor(uint const & level, uint const & i, uint const & j, uint const & dir) {
+        size_t & get_neighbor(size_t const & level, size_t const & i, size_t const & j, size_t const & dir) {
             switch(dir) {
                 case(qmc::down):
                     return stage2_[level][(i + 1) % H_][j];
@@ -138,22 +138,22 @@ namespace perimeter {
         
         void grow_partial(double steps_in = 1) {
             steps_in -= 2;
-            uint steps = std::abs(steps_in) + 1;
-            uint grow_count = 0;
-            uint max_grow = std::round(std::abs(steps_in) * H_);
-            const uint just_grown = 1000;
+            size_t steps = std::abs(steps_in) + 1;
+            size_t grow_count = 0;
+            size_t max_grow = std::round(std::abs(steps_in) * H_);
+            const size_t just_grown = 1000;
             grow_level_ += steps;
-            for(uint level = 0; level < N_; ++level) {
+            for(size_t level = 0; level < N_; ++level) {
                 grow_count = 0;
-                for(uint k = 0; k < steps; ++k) {
-                    for(uint j = 0; j < L_; ++j) {
-                        for(uint i = 0; i < H_; ++i) {
-                            uint i_eff = steps_in < 0 ? H_ - i - 1: i;
-                            uint j_eff = steps_in < 0 ? L_ - j - 1: j;
-                            uint grow_factor = stage2_[level][i_eff][j_eff];
+                for(size_t k = 0; k < steps; ++k) {
+                    for(size_t j = 0; j < L_; ++j) {
+                        for(size_t i = 0; i < H_; ++i) {
+                            size_t i_eff = steps_in < 0 ? H_ - i - 1: i;
+                            size_t j_eff = steps_in < 0 ? L_ - j - 1: j;
+                            size_t grow_factor = stage2_[level][i_eff][j_eff];
                             if(grow_factor > 0 and grow_factor < just_grown) {
                                 std::for_each(grow_dir_.begin(), grow_dir_.end(), 
-                                    [&](uint const & dir) {
+                                    [&](size_t const & dir) {
                                         if(get_neighbor(level, i_eff, j_eff, dir) == 0){
                                             if(grow_count++ < max_grow) {
                                                 if(steps_in < 0)
@@ -167,8 +167,8 @@ namespace perimeter {
                             }
                         }
                     }
-                    for(uint i = 0; i < H_; ++i) {
-                        for(uint j = 0; j < L_; ++j) {
+                    for(size_t i = 0; i < H_; ++i) {
+                        for(size_t j = 0; j < L_; ++j) {
                             if(stage2_[level][i][j] >= just_grown)
                                 stage2_[level][i][j] -= just_grown;
                         }
@@ -179,17 +179,17 @@ namespace perimeter {
         
         void grow(int steps_in = 1) {
             steps_in -= 2;//TODO: remove later
-            const uint just_grown = 1000;
-            uint steps = std::abs(steps_in);
+            const size_t just_grown = 1000;
+            size_t steps = std::abs(steps_in);
             grow_level_ += steps;
-            for(uint level = 0; level < N_; ++level) {
-                for(uint k = 0; k < steps; ++k) {
-                    for(uint i = 0; i < H_; ++i) {
-                        for(uint j = 0; j < L_; ++j) {
-                            uint grow_factor = stage2_[level][i][j];
+            for(size_t level = 0; level < N_; ++level) {
+                for(size_t k = 0; k < steps; ++k) {
+                    for(size_t i = 0; i < H_; ++i) {
+                        for(size_t j = 0; j < L_; ++j) {
+                            size_t grow_factor = stage2_[level][i][j];
                             if(grow_factor > 0 and grow_factor < just_grown) {
                                 std::for_each(grow_dir_.begin(), grow_dir_.end(), 
-                                    [&](uint const & dir) {
+                                    [&](size_t const & dir) {
                                         if(get_neighbor(level, i, j, dir) == 0){
                                             if(steps_in < 0)
                                                 stage2_[level][i][j] = just_grown;
@@ -201,8 +201,8 @@ namespace perimeter {
                             }
                         }
                     }
-                    for(uint i = 0; i < H_; ++i) {
-                        for(uint j = 0; j < L_; ++j) {
+                    for(size_t i = 0; i < H_; ++i) {
+                        for(size_t j = 0; j < L_; ++j) {
                             if(stage2_[level][i][j] >= just_grown)
                                 stage2_[level][i][j] -= just_grown;
                         }
@@ -215,8 +215,8 @@ namespace perimeter {
             convert_2_to_1();
             
             std::ofstream os(filename);
-            for(uint n = 0; n < N_; ++n) {
-                for(uint i = 0; i < H_; ++i) {
+            for(size_t n = 0; n < N_; ++n) {
+                for(size_t i = 0; i < H_; ++i) {
                     os << stage1_[n][i] << std::endl;
                 }
                 if(n != N_ - 1)
@@ -226,10 +226,10 @@ namespace perimeter {
         }
     private:
         void convert_1_to_2() {
-            stage2_ = std::vector<std::vector<std::vector<uint>>>(N_, std::vector<std::vector<uint>>(H_, std::vector<uint>(L_, 0)));
-            for(uint n = 0; n < N_; ++n) {
-                for(uint i = 0; i < H_; ++i) {
-                    for(uint j = 0; j < L_; ++j) {
+            stage2_ = std::vector<std::vector<std::vector<size_t>>>(N_, std::vector<std::vector<size_t>>(H_, std::vector<size_t>(L_, 0)));
+            for(size_t n = 0; n < N_; ++n) {
+                for(size_t i = 0; i < H_; ++i) {
+                    for(size_t j = 0; j < L_; ++j) {
                         stage2_[n][i][j] = (stage1_[n][i][2*j] - '0');
                     }
                 }
@@ -237,31 +237,31 @@ namespace perimeter {
         }
         void convert_2_to_1() {
             stage1_ =  std::vector<std::vector<std::string>>(N_, std::vector<std::string>(H_, ""));
-            for(uint n = 0; n < N_; ++n) {
-                for(uint i = 0; i < H_; ++i) {
-                    for(uint j = 0; j < L_; ++j) {
+            for(size_t n = 0; n < N_; ++n) {
+                for(size_t i = 0; i < H_; ++i) {
+                    for(size_t j = 0; j < L_; ++j) {
                         stage1_[n][i] += '0' + (*this)(n, i, j);
                         stage1_[n][i] += " ";
                     }
                 }
             }
         }
-        uint count_sites(std::string const & in) const {
-            uint res(0);
-            for(uint i = 0; i < in.size(); ++i) {
+        size_t count_sites(std::string const & in) const {
+            size_t res(0);
+            for(size_t i = 0; i < in.size(); ++i) {
                 if(in[i] >= '0' and in[i] <= '9')
                     ++res;
             }
             return res;
         }
     private:
-        uint H_;
-        uint L_;
-        uint N_;
+        size_t H_;
+        size_t L_;
+        size_t N_;
         std::vector<std::vector<std::string>> stage1_;
-        std::vector<std::vector<std::vector<uint>>> stage2_;
+        std::vector<std::vector<std::vector<size_t>>> stage2_;
         std::vector<bond_type> grow_dir_;
-        uint grow_level_;
+        size_t grow_level_;
     };
 }//end namespace perimeter
